new file:   qbasegraphicsview.cpp 	new file:   qbasegraphicsview.h
Another refactor to tidy up the remaining issues.

Refactored and migrated the QBaseGraphicsView to its own file in order to clean up the code.

Finished up the full migration of the frame slider to the QGraphicsView to aid in future development for thumbnails and in / out markers for example.

Refactored some of the lower level ffSequence object virtual function calls to make GUI management easier.

The export option is still not finished, and currently only saves the luma (Y') channel to EXR format. This was largely to test OpenImageIO integration. Needs to have the full frame range output. In finished form it will permit bounded frame ranges to be export.

Next phase will likely involve migrating to the first conversion to sRGB.