Refactored to tidy up evolving design needs. Attempted to clarify the planes enums to avoid compiler warnings in the unhandled switch / case stanzas.
Added interface feedback on frame changes and asserted that repainting happens correctly.

Synchronized all of the UI with the new Export options UI elements. Synchronized the Sidebar options to correctly be disabled and enabled according to validity of file. Export bounding of in / out points established.

Still must hook up the values for the in / out points to the Exporting function. Getting close to being useful...