new file:   TODO.txt 	new file:   qgraphicstrimslider.cpp 	new file:   qgraphicstrimslider.h
Added a customized trim control frame scroller that better reflects the needs of this application.

Hooked up the frame scroller with in and out configuration that touches the underlying ExportDetails. Still outstanding are the sidebar interface elements. Irritating event sequencing that trumps the freshly "Loaded file..." message with the in out trim message.